Move podStoreAdapter into podstoreadapter.go
