+++ contrib/mesos/pkg/scheduler/plugin.go
@@ -360,42 +360,3 @@ func createAllPodsLW(cl *client.Client) *cache.ListWatch {
 	return cache.NewListWatchFromClient(cl, "pods", api.NamespaceAll, parseSelectorOrDie(""))
 }
 
-// Consumes *api.Pod, produces *Pod; the k8s reflector wants to push *api.Pod
-// objects at us, but we want to store more flexible (Pod) type defined in
-// this package. The adapter implementation facilitates this. It's a little
-// hackish since the object type going in is different than the object type
-// coming out -- you've been warned.
-type podStoreAdapter struct {
-	queue.FIFO
-}
-
-func (psa *podStoreAdapter) Add(obj interface{}) error {
-	pod := obj.(*api.Pod)
-	return psa.FIFO.Add(&queuer.Pod{Pod: pod})
-}
-
-func (psa *podStoreAdapter) Update(obj interface{}) error {
-	pod := obj.(*api.Pod)
-	return psa.FIFO.Update(&queuer.Pod{Pod: pod})
-}
-
-func (psa *podStoreAdapter) Delete(obj interface{}) error {
-	pod := obj.(*api.Pod)
-	return psa.FIFO.Delete(&queuer.Pod{Pod: pod})
-}
-
-func (psa *podStoreAdapter) Get(obj interface{}) (interface{}, bool, error) {
-	pod := obj.(*api.Pod)
-	return psa.FIFO.Get(&queuer.Pod{Pod: pod})
-}
-
-// Replace will delete the contents of the store, using instead the
-// given map. This store implementation does NOT take ownership of the map.
-func (psa *podStoreAdapter) Replace(objs []interface{}, resourceVersion string) error {
-	newobjs := make([]interface{}, len(objs))
-	for i, v := range objs {
-		pod := v.(*api.Pod)
-		newobjs[i] = &queuer.Pod{Pod: pod}
-	}
-	return psa.FIFO.Replace(newobjs, resourceVersion)
-}
